Move job's status to explore job-status
